+++ Demo/src/main/java/com/braintreepayments/demo/MainActivity.java
@@ -116,7 +116,6 @@ public class MainActivity extends BaseActivity implements PaymentMethodNonceCrea
     public void launchDropIn(View v) {
         DropInRequest dropInRequest = new DropInRequest()
                 .clientToken(mAuthorization)
-                .amount("1.00")
                 .requestThreeDSecureVerification(Settings.isThreeDSecureEnabled(this))
                 .collectDeviceData(Settings.shouldCollectDeviceData(this))
                 .googlePaymentRequest(getGooglePaymentRequest())